pi/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
demucs==4.0.1
basic-pitch==0.3.2
boto3==1.34.0
python-dotenv==1.0.0
numpy<2
